requirements.txt
requests
beautifulsoup4
httpx
python-dotenv
google-genai>=0.3.0
pandas
